requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
spacy==3.7.2
python-dotenv==1.0.0
werkzeug==3.0.1
requests==2.31.0
gunicorn==21.2.0
